Documentation and code standardization (#76)

+++ public/Disable-DbsCEIP.ps1
@@ -12,6 +12,12 @@ function Disable-DbsCEIP {
     .PARAMETER Credential
         Credential object used to connect to the computer as a different user.
 
+    .PARAMETER WhatIf
+        If this switch is enabled, no actions are performed but informational messages will be displayed that explain what would happen if the command were to run
+
+    .PARAMETER Confirm
+        If this switch is enabled, you will be prompted for confirmation before executing any operations that change state
+
     .PARAMETER EnableException
         By default, when something goes wrong we try to catch it, interpret it and give you a friendly warning message.
         This avoids overwhelming you with "sea of red" exceptions, but is inconvenient because it basically Gets advanced scripting.
@@ -32,8 +38,8 @@ function Disable-DbsCEIP {
         PS C:\> Disable-DbsCEIP -ComputerName sql2016, sql2017, sql2012 -Credential ad\altdba
 
         Disables all instances of CEIP on sql2016, sql2017 and sql2012 using alternative Windows credentials
-#>
-    [CmdletBinding()]
+    #>
+    [CmdletBinding(SupportsShouldProcess, ConfirmImpact = "Medium")]
     param (
         [parameter(Mandatory, ValueFromPipeline)]
         [DbaInstanceParameter[]]$ComputerName,
@@ -41,28 +47,30 @@ function Disable-DbsCEIP {
         [switch]$EnableException
     )
     begin {
-        $PSDefaultParameterValues['*:EnableException'] = $true
-        $PSDefaultParameterValues['Stop-PSFFunction:EnableException'] = $EnableException
-        $PSDefaultParameterValues['*:Credential'] = $Credential
+        . "$script:ModuleRoot\private\Set-Defaults.ps1"
     }
     process {
         foreach ($computer in $ComputerName.ComputerName) {
-            try {
-                # thanks to https://blog.dbi-services.com/sql-server-tips-deactivate-the-customer-experience-improvement-program-ceip/
-                Invoke-PSFCommand -ComputerName $computer -ScriptBlock {
-                    $services = Get-Service | Where-Object Name -Like "*TELEMETRY*"
-                    $services | Stop-Service
-                    $services | Set-Service -StartupType Disabled
-                    $keys = Get-ChildItem 'HKLM:\SOFTWARE\Microsoft\Microsoft SQL Server', 'HKLM:\SOFTWARE\Wow6432Node\Microsoft\Microsoft SQL Server' -Recurse -ErrorAction Stop | Where-Object -Property Property -eq 'EnableErrorReporting'
-                    foreach ($key in $keys) {
-                        $null = $key | Set-ItemProperty -Name EnableErrorReporting -Value 0
-                        $null = $key | Set-ItemProperty -Name CustomerFeedback -Value 0
+            if ((Get-DbsCEIP -Computer $computer).Enabled) {
+                if ($PSCmdlet.ShouldProcess($computer, "Disabling telemetry")) {
+                    try {
+                        # thanks to https://blog.dbi-services.com/sql-server-tips-deactivate-the-customer-experience-improvement-program-ceip/
+                        Invoke-PSFCommand -ComputerName $computer -ScriptBlock {
+                            $services = Get-Service | Where-Object Name -Like "*TELEMETRY*"
+                            $services | Stop-Service
+                            $services | Set-Service -StartupType Disabled
+                            $keys = Get-ChildItem 'HKLM:\SOFTWARE\Microsoft\Microsoft SQL Server', 'HKLM:\SOFTWARE\Wow6432Node\Microsoft\Microsoft SQL Server' -Recurse -ErrorAction Stop | Where-Object -Property Property -eq 'EnableErrorReporting'
+                            foreach ($key in $keys) {
+                                $null = $key | Set-ItemProperty -Name EnableErrorReporting -Value 0
+                                $null = $key | Set-ItemProperty -Name CustomerFeedback -Value 0
+                            }
+                        }
+                    } catch {
+                        Stop-PSFFunction -Message "Failure on $computer" -ErrorRecord $_ -Continue
                     }
                 }
-                Get-DbsCEIP -Computer $computer
-            } catch {
-                Stop-PSFFunction -Message "Failure on $computer" -ErrorRecord $_ -Continue
             }
+            Get-DbsCEIP -Computer $computer
         }
     }
 }
